Update tests for getResult for basics
