+++ misc/ls-extensions/modules/bal-shell-service/src/test/java/io/ballerina/shell/service/test/getresult/AbstractGetResultTest.java
@@ -15,7 +15,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package io.ballerina.shell.service.test.getgesult;
+package io.ballerina.shell.service.test.getresult;
 
 import io.ballerina.shell.service.BalShellGetResultRequest;
 import io.ballerina.shell.service.BalShellGetResultResponse;
