+++ misc/ls-extensions/modules/bal-shell-service/src/test/java/io/ballerina/shell/service/test/getresult/GetResultTestCase.java
@@ -15,21 +15,24 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package io.ballerina.shell.service.test.getgesult;
+package io.ballerina.shell.service.test.getresult;
 
-import org.testng.annotations.Test;
-
-import java.io.IOException;
-import java.util.concurrent.ExecutionException;
+import io.ballerina.shell.service.BalShellGetResultResponse;
 
 /**
- * getResult End point tests for simple arithmetics.
+ * Class to represent GetResultTestCase json.
  *
  * @since 2201.1.1
  */
-public class SimpleArithmeticGetResultTest extends AbstractGetResultTest {
-    @Test(description = "Test with simple arithmetic")
-    public void testSimpleArithmetic() throws ExecutionException, IOException, InterruptedException {
-        runGetResultTest("simple_arithmetic.json");
+public class GetResultTestCase {
+    private String source;
+    private BalShellGetResultResponse result;
+
+    public String getSource() {
+        return source;
+    }
+
+    public BalShellGetResultResponse getResult() {
+        return result;
     }
 }
