+++ misc/ls-extensions/modules/bal-shell-service/src/test/resources/testcases/getResult/basic.arrays.json
@@ -1,0 +1,94 @@
+[
+  {
+    "source": "int[] array1 = [1, 2, 3];",
+    "result": {
+      "errors":[],
+      "diagnostics":[],
+      "metaInfo":{"definedVars":["array1"],"moduleDclns":[]},
+      "consoleOut":""
+    }
+  },
+  {
+    "source": "array1",
+    "result": {
+      "shellValue": {
+        "value":"[1, 2, 3]",
+        "mimeType":"ballerina-notebook/json-view",
+        "type":"int[]"
+      },
+      "errors":[],
+      "diagnostics":[],
+      "metaInfo":{"definedVars":[],"moduleDclns":[]},
+      "consoleOut":""
+    }
+  },
+  {
+    "source": "byte[] byteArray1 = [5, 24, 56, 243];byte[] byteArray2 = base16 `aeeecdefabcd12345567888822`;",
+    "result": {
+      "errors":[],
+      "diagnostics":[],
+      "metaInfo":{"definedVars":["byteArray1", "byteArray2"],"moduleDclns":[]},
+      "consoleOut":""
+    }
+  },
+  {
+    "source": "byteArray1",
+    "result": {
+      "shellValue": {
+        "value":"[5, 24, 56, 243]",
+        "mimeType":"ballerina-notebook/json-view",
+        "type":"byte[]"
+      },
+      "errors":[],
+      "diagnostics":[],
+      "metaInfo":{"definedVars":[],"moduleDclns":[]},
+      "consoleOut":""
+    }
+  },
+  {
+    "source": "byteArray2",
+    "result": {
+      "shellValue": {
+        "value":"[174, 238, 205, 239, 171, 205, 18, 52, 85, 103, 136, 136, 34]",
+        "mimeType":"ballerina-notebook/json-view",
+        "type":"byte[]"
+      },
+      "errors":[],
+      "diagnostics":[],
+      "metaInfo":{"definedVars":[],"moduleDclns":[]},
+      "consoleOut":""
+    }
+  },
+  {
+    "source": "anydata[] data = [12, \"hello\", byteArray2, ()]",
+    "result": {
+      "errors":[],
+      "diagnostics":[],
+      "metaInfo":{"definedVars":["data"],"moduleDclns":[]},
+      "consoleOut":""
+    }
+  },
+  {
+    "source": "data",
+    "result": {
+      "shellValue": {
+        "value":"[12, \"hello\", [174, 238, 205, 239, 171, 205, 18, 52, 85, 103, 136, 136, 34], null]",
+        "mimeType":"ballerina-notebook/json-view",
+        "type":"anydata[]"
+      },
+      "errors":[],
+      "diagnostics":[],
+      "metaInfo":{"definedVars":[],"moduleDclns":[]},
+      "consoleOut":""
+    }
+  },
+  {
+    "source": "const int[] constArray = [4, 5 , 6]",
+    "result": {
+      "errors":["Invocation of the snippet failed."],
+      "diagnostics":["error: constant declaration not yet supported for type \u0027int[]\u0027\n\tconst int[] constArray \u003d [4, 5 , 6];\n\t      ^---^","error: expression is not a constant expression\n\tconst int[] constArray \u003d [4, 5 , 6];\n\t                         ^--------^","Compilation aborted due to errors."],
+      "metaInfo":{"definedVars":[],"moduleDclns":[]},
+      "consoleOut":""
+    }
+  }
+]
