--- misc/ls-extensions/modules/bal-shell-service/src/test/resources/testcases/getResult/basic_variables.json
@@ -1,60 +1,0 @@
-[
-  {
-    "source": "int total = 99;",
-    "result": {
-      "errors":[],
-      "diagnostics":[],
-      "metaInfo":{"definedVars":["total"],"moduleDclns":[]},
-      "consoleOut":""
-    }
-  },
-  {
-    "source": "string s = \"Hello world\"",
-    "result": {
-      "errors":[],
-      "diagnostics":[],
-      "metaInfo":{"definedVars":["s"],"moduleDclns":[]},
-      "consoleOut":""
-    }
-  },
-  {
-    "source": "total",
-    "result": {
-      "shellValue": {
-        "value":"99",
-        "mimeType":"plain/text",
-        "type":"int"
-      },
-      "errors":[],
-      "diagnostics":[],
-      "metaInfo":{"definedVars":[],"moduleDclns":[]}
-    }
-  },
-  {
-    "source": "total*2",
-    "result": {
-      "shellValue": {
-        "value":"198",
-        "mimeType":"plain/text",
-        "type":"int"
-      },
-      "errors":[],
-      "diagnostics":[],
-      "metaInfo":{"definedVars":[],"moduleDclns":[]},
-      "consoleOut":""
-    }
-  },
-  {
-    "source": "s",
-    "result": {
-      "shellValue": {
-        "value":"\"Hello world\"",
-        "mimeType":"plain/text",
-        "type":"string"
-      },
-      "errors":[],
-      "diagnostics":[],
-      "metaInfo":{"definedVars":[],"moduleDclns":[]}
-    }
-  }
-]
