--- misc/ls-extensions/modules/bal-shell-service/src/test/resources/testcases/getResult/functions.json
@@ -1,34 +1,0 @@
-[
-  {
-    "source": "int x = 10;",
-    "result": {
-      "errors":[],
-      "diagnostics":[],
-      "metaInfo":{"definedVars":["x"],"moduleDclns":[]},
-      "consoleOut":""
-    }
-  },
-  {
-    "source": "function multiply(int m, int n) returns int { return m * n;}",
-    "result": {
-      "errors":[],
-      "diagnostics":[],
-      "metaInfo":{"definedVars":[],"moduleDclns":["multiply"]},
-      "consoleOut":""
-    }
-  },
-  {
-    "source": "multiply(x,5)",
-    "result": {
-      "shellValue": {
-        "value":"50",
-        "mimeType":"plain/text",
-        "type":"int"
-      },
-      "errors":[],
-      "diagnostics":[],
-      "metaInfo":{"definedVars":[],"moduleDclns":[]},
-      "consoleOut":""
-    }
-  }
-]
